move tests to large and use page size to speed up
